remove standalone metastore server, integrate it into server master
