+++ cmd/executor/example.toml
@@ -2,5 +2,4 @@ join = "127.0.0.1:10240"
 worker-addr = "127.0.0.1:10241"
 keepalive-ttl = "10s"
 keepalive-interval = "500ms"
-meta-endpoints = "127.0.0.1:10340"
 session-ttl = 10
