--- cmd/metastore/example.toml
@@ -1,4 +1,0 @@
-addr = "http://127.0.0.1:10340"
-[etcd]
-name = "metastore-1"
-data-dir = "/tmp/df/metastore/discovery"
