--- executor/meta.go
@@ -1,1 +1,0 @@
-package executor
